requirements.txt
Django==5.0
djangorestframework==3.14.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-cors-headers==4.3.1
drf-yasg==1.21.7
python-dotenv==1.0.0
